python/email-order-agent/requirements.txt
azure-functions>=1.17.0
openai>=1.0.0
requests>=2.31.0
azure-identity>=1.15.0
